safely removed CRASparkMax and Util
