+++ src/main/java/frc/robot/subsystems/Conveyor.java
@@ -5,13 +5,24 @@ import edu.wpi.first.wpilibj2.command.RunCommand;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
 import java.util.function.BooleanSupplier;
-import CRA.CRASparkMax;
+
+import com.revrobotics.spark.SparkBase;
+import com.revrobotics.spark.SparkLowLevel.MotorType;
+import com.revrobotics.spark.config.SparkBaseConfig;
+import com.revrobotics.spark.config.SparkMaxConfig;
+import com.revrobotics.spark.SparkMax;
 
 public class Conveyor extends SubsystemBase {
-    CRASparkMax motor;
+    private final SparkMax motor = new SparkMax(Constants.Conveyor.MOTOR_CAN_ID, MotorType.kBrushless);
+    private final SparkMaxConfig config = new SparkMaxConfig();
 
     public Conveyor() {
-        motor = new CRASparkMax(Constants.Conveyor.MOTOR_CAN_ID, false, false);
+        config.idleMode(SparkBaseConfig.IdleMode.kCoast);
+        config.smartCurrentLimit(Constants.Conveyor.CURRENT_LIMIT);
+        motor.configure(config,
+            SparkBase.ResetMode.kResetSafeParameters, 
+            SparkBase.PersistMode.kPersistParameters
+        );
     }
     
     public Command clearCoral(BooleanSupplier hasCoralToClear) {
